Move callback via to evtchn_op
When the guest uses HVMOP_set_param to set the callback via,
Xen calls hvm_assert_evtchn_irq, which ultimately results
in a vcpu_kick/interrupt injection.

We don't have a scheduler, so for now our vcpu_kick and
injection are one in the same. We queue an interrupt
at the callback vector, after setting the
evtchn_upcall_pending in the shared vcpu_info.

The evtchn_pending_sel field is set before the callback is
set_param'd in evtchn_op::setup_ports

Signed-off-by: Connor Davis <davisc@ainfosec.com>