Add ioctl_mmap and ioctl_munmap
IOCTL_MMAP
----------
Allows userspace to allocate bytes from the kernel
with the promise that the memory will never be paged out of
memory and will already be paged in upon successful return.

That last bit is necessary so that the entries are valid when
the hypervisor walks the guest's page tables.

IOCTL_MUNMAP
------------
Unmaps the provided mmap

Signed-off-by: Connor Davis <davisc@ainfosec.com>